git-svn-id: http://agilefant.googlecode.com/svn/trunk@3668 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ test/fi/hut/soberit/agilefant/web/ProjectActionTest.java
@@ -127,7 +127,7 @@ public class ProjectActionTest {
         replayAll();
         projectAction.setConfirmationString("yes");
         assertEquals(Action.SUCCESS, projectAction.delete());
-        assertEquals(123, projectAction.getProductId().intValue());
+        assertEquals(123, projectAction.getProductId());
         verifyAll();    
     }
